Extract transaction observer
